refactor(bridge): split pymethods by concern, decompose init_planner and attach_serial
bridge.rs 3783 -> 1130 lines (MCU lifecycle + struct + shared helpers).
New per-concern pymethods blocks via pyo3 multiple-pymethods: servo,
passthrough, planner_api, telemetry, homing_api. init_planner now reads
as seven named steps; attach_serial as three. Python-visible API
unchanged.
