+++ rust/motion-engine/Cargo.toml
@@ -14,7 +14,7 @@ name = "_motion_engine"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-pyo3 = { version = "0.29", features = ["abi3-py39"] }
+pyo3 = { version = "0.29", features = ["abi3-py39", "multiple-pymethods"] }
 host-rt = { path = "../host-rt", features = ["python-engine"] }
 mcu-protocol = { path = "../mcu-protocol" }
 trajectory = { path = "../trajectory" }
